Implemented new contraction check.
The base case of the contraction algorithm checks that there is a
unique top dimensional generator in the colimit diagram. This is
not sufficient to avoid ill-typed diagrams, so an additional type
checking phase is required.

To potentially remove the need for this type checking phase, the
check performed in the base case was extended: not only needs the
top dimensional generator be unique, but all instances of this
generator in the colimit diagram have to be connected via identity
maps of that generator.